+++ src/storage/index/hash/key_value_hash/key_value_hash_bucket.h
@@ -37,7 +37,7 @@ private:
 
     uint32_t   const key_size;
     uint32_t   const value_size;
-    uint32_t*  const tuple_count;
+    uint64_t*  const tuple_count;  // we use 64 bits for alignment
     char*      const tuples;
 };
